kafka_producer/requirements.txt
kafka-python==2.0.2
aiokafka==0.7.2
pydantic>=2.0.0
